Move proxy to root
